// typings/statcrew/football/team.d.ts
interface IFumbleStats {
  $: {
    no: number;
    lost: number;
  };
}

interface IMiscStats {
  $: {
    yds: number;
    top: string; // time
    ona: number;
    onm: number;
    ptsto: number;
  };
}

interface IRedzoneStats {
  $: {
    att: number;
    scores: number;
    points: number;
    tdrush: number;
    tdpass: number;
    fgmade: number;
    endfga: number;
    enddowns: number;
    endint: number;
    endfumb: number;
    endhalf: number;
    endgame: number;
  };
}

interface IRushStats {
  $: {
    att: number;
    yds: number;
    gain: number;
    loss: number;
    td: number;
    long: number;
  };
}

interface IPassStats {
  $: {
    comp: number;
    att: number;
    int: number;
    yds: number;
    td: number;
    long: number;
    sacks: number;
    sackyds: number;
  };
}

interface IReceivingStats {
  $: {
    no: number;
    yds: number;
    td: number;
    long: number;
  };
}

interface IPuntStats {
  $: {
    no: number;
    yds: number;
    long: number;
    avg?: number;
    blkd: number;
    tb: number;
    fc: number;
    plus50: number;
    inside20: number;
  };
}

interface IKickoffStats {
  $: {
    no: number;
    yds: number;
    ob: number;
    tb: number;
  };
}

interface IFieldgoalStats {
  $: {
    made: number;
    att: number;
    long: number;
    blkd: number;
  };
}

interface IPointAfterStats {
  $: {
    kickatt: number;
    kickmade: number;
    passatt?: number;
    passmade?: number;
    rushatt?: number;
    rushmade?: number;
  };
}

interface IDefenseStats {
  $: {
    tackua: number;
    tacka: number;
    tot_tack?: number;
    tflua?: number;
    tfla?: number;
    tflyds?: number;
    sacks?: number;
    sackyds?: number;
    sackua?: number;
    sacka?: number;
    brup?: number;
    ff?: number;
    fr?: number;
    fryds?: number;
    int?: number;
    intyds?: number;
    qbh?: number;
    blkd?: number;
  };
}

interface ICommonReturnStats {
  $: {
    no: number;
    yds: number;
    td: number;
    long: number;
  };
}

interface IScoringStats {
  $: {
    fg?: number;
    td?: number;
    patkick?: number;
  };
}
// ...
interface IPlayer {
  $: {
    name: string;
    shortname?: string;
    checkname: string;
    uni: string | number;
    playerid?: string;
    class: "FR" | "SO" | "JR" | "SR";
    gp: number;
    gs?: number;
    opos?: string;
    dpos?: string;
    code: string | number;
  };
  rcv?: IReceivingStats[];
  rush?: IRushStats[];
  pass?: IPassStats[];
  defense?: IDefenseStats[];
  kr?: ICommonReturnStats[];
  fumbles?: IFumbleStats[];
  ir?: ICommonReturnStats[];
  scoring?: IScoringStats[];
  ko?: IKickoffStats[];
  fg?: IFieldgoalStats[];
  pat?: IPointAfterStats[];
  punt?: IPuntStats[];
  pr?: ICommonReturnStats[];
  fr?: ICommonReturnStats[];
  misc?: IMiscStats[];
}